Remove Create New Incident doc tool
This tool never worked because we're using native GitHub pages instead of running jekyll ourselves in GitHub actions.
As a result, CSP prevents the tool from working.

As part of this commit, I'm deleting the `apiKey` used which, had CSP worked, granted rights to `*.mozilla.org/*` sites (calls with that referer) to interact with the Google Drive API. This API Key was provisioned in the `mozilla.com` GCP org in the `Security Incident Doc Creator` project (`security-incident-doc-creator`).

If we want to use this in the future, or do anything more complex than the static web pages we have now, we will need to switch infosec.mozilla.org to a jekyll site which is generated using GitHub Actions.